+++ src/components/ArrayField.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import useArrayField from '../hooks/useArrayField';
 
-const ArrayField = ({ field, children, initialValue, validate }) => {
+const ArrayField = ({ children, ...props }) => {
 
-  const arrayField = useArrayField({ field, initialValue, validate});
+  const arrayField = useArrayField(props);
 
   return children(arrayField);
 };